+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.2.3
+
+* Added `closure(..)` function to create closures from string, issue #273.
+* Fixed task replication bug in `set_closure(..)` function, issue #274.
+* Fixed task replication bug in `set_owner(..)` function, issue #275.
+
 # v1.2.2
 
 * Fixed storing migration change for *non-collection* scopes, issue #269.